review-fix: Dynamically add WS port from CLI to trusted
