+++ packages/fether-electron/src/main/index.js
@@ -9,13 +9,12 @@ import { killParity } from '@parity/electron';
 
 import Pino from './app/utils/pino';
 import FetherApp from './app';
-import { TRUSTED_LOOPBACK_PREFIX, TRUSTED_URLS } from './app/constants';
+import { SECURITY_OPTIONS } from './app/options/config';
 import fetherAppOptions from './app/options';
-// import { customWsPort } from './app/cli';
-import cli from './app/cli';
 
-const { app, shell } = electron;
 const pino = Pino();
+const { app, shell } = electron;
+const { TRUSTED_URLS } = SECURITY_OPTIONS.network;
 
 let withTaskbar = process.env.TASKBAR !== 'false';
 
@@ -24,8 +23,6 @@ pino.info('Process type: ', process.type);
 pino.info('Process ID: ', process.pid);
 pino.info('Process args: ', process.argv);
 pino.info('Electron version: ', process.versions['electron']);
-// pino.info('customWsPort', customWsPort);
-pino.info('cli.wsPort: ', cli.wsPort);
 
 // Disable gpu acceleration on linux
 // https://github.com/parity-js/fether/issues/85
@@ -118,12 +115,7 @@ app.on('web-contents-created', (eventOuter, win) => {
       parsedUrl.href
     );
 
-    if (
-      !TRUSTED_URLS.includes(parsedUrl.href) ||
-      // customWsPort && parsedUrl.href !== `${TRUSTED_LOOPBACK_PREFIX}:${customWsPort}`
-      (cli.wsPort &&
-        parsedUrl.href !== `${TRUSTED_LOOPBACK_PREFIX}:${cli.wsPort}`)
-    ) {
+    if (!TRUSTED_URLS.includes(parsedUrl.href)) {
       pino.info(
         'Unable to navigate to untrusted content url due to will-navigate listener: ',
         parsedUrl.href
@@ -154,12 +146,7 @@ app.on('web-contents-created', (eventOuter, win) => {
         parsedUrl.href
       );
 
-      if (
-        !TRUSTED_URLS.includes(parsedUrl.href) ||
-        // customWsPort && parsedUrl.href !== `${TRUSTED_LOOPBACK_PREFIX}:${customWsPort}`
-        (cli.wsPort &&
-          parsedUrl.href !== `${TRUSTED_LOOPBACK_PREFIX}:${cli.wsPort}`)
-      ) {
+      if (!TRUSTED_URLS.includes(parsedUrl.href)) {
         pino.info(
           'Unable to open new window with untrusted content url due to new-window listener: ',
           parsedUrl.href
